Documentation updated
